Make miniscene 360images scale with zoom
When the user zooms in/out on the miniscene, the spheres grow/shrink to
take up the same constant amount of screen space. There is a max size
that prevents them from turning into a jumble when the user zooms out
too far.

The miniscene's camera is also orthogonal now. This stops the spheres
from moving around in perspective since they are in front of the
background.

This commit matches a commit in cpms-web.