Add material, rotate to face camera
Neighbor markers now rotate vertically to face the camera, not just
horizontally. They also have a specified material.

The 360image sphere renders on top of other objects regardless of
distance, which means it renders on top of the neighbor markers.

The neighbor markers now use a nearly flat 3D object, but this looks
strange while moving from image to image so it would probably be better
to use 2D sprites. I am just committing this for future reference.